+++ packages/expandable-input/package/src/index.ts
@@ -14,6 +14,7 @@ export * from './lib/exp-icon-action.directive';
 export * from './lib/exp-icon-close.directive';
 export * from './lib/exp-icon-open.directive';
 export * from './lib/exp-input.directive';
+export * from './lib/expandable-input-base';
 export * from './lib/expandable-input.component';
 export * from './lib/expandable-input.module';
 export * from './lib/inputs-manager.service';