+++ packages/expandable-input/package/src/lib/exp-icon-open.directive.ts
@@ -1,7 +1,9 @@
-import { Directive } from '@angular/core';
+import { Directive, inject, TemplateRef } from '@angular/core';
 
 @Directive({
   selector: '[ngsExpIconOpen]',
   standalone: true,
 })
-export class ExpIconOpenDirective {}
+export class ExpIconOpenDirective {
+  templateRef = inject(TemplateRef);
+}